+++ packages/local_auth/local_auth_ios/lib/local_auth_ios.dart
@@ -98,8 +98,8 @@ class LocalAuthIOS extends LocalAuthPlatform {
   Future<bool> stopAuthentication() async => false;
 
   @override
-  Future<bool> isValidBiometricAuthorized() async =>
-      _api.isValidBiometricAuthorized();
+  Future<void> validateBiometricAuthorized() async =>
+      _api.validateBiometricAuthorized();
 
   AuthStrings _pigeonStringsFromAuthMessages(
       String localizedReason, Iterable<AuthMessages> messagesList) {