+++ packages/local_auth/local_auth_ios/pigeons/messages.dart
@@ -97,8 +97,8 @@ abstract class LocalAuthApi {
   /// Returns true if this device supports authentication.
   bool isDeviceSupported();
 
-  /// Returns false if user change biometric value you must to authorize the new biometric.
-  bool isValidBiometricAuthorized();
+  /// Validate if user change biometric value you must to authorize the new biometric.
+  void validateBiometricAuthorized();
 
   /// Returns true if this device can support biometric authentication, whether
   /// any biometrics are enrolled or not.