+++ packages/local_auth/local_auth_platform_interface/lib/default_method_channel_platform.dart
@@ -83,7 +83,6 @@ class DefaultLocalAuthPlatform extends LocalAuthPlatform {
       await _channel.invokeMethod<bool>('stopAuthentication') ?? false;
 
   @override
-  Future<bool> isValidBiometricAuthorized() async =>
-      (await _channel.invokeMethod<bool>('isValidBiometricAuthorized')) ??
-      false;
+  Future<void> validateBiometricAuthorized() =>
+      _channel.invokeMethod('validateBiometricAuthorized');
 }