+++ packages/local_auth/local_auth_ios/CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ## 1.1.4
 
-- Added isValidBiometricAuthorized, should be used when you want to authorize register device biometric.
+- Added validateBiometricAuthorized, should be used when you want to authorize register device biometric.
 - Change authorize method to invalidate when user change biometric device.
 
 ## 1.1.3